DHT22, Temperature, Thermometer, Pin, macros, debug messages & cleanup. Implemented and tested a controller class for the DHT22 digital humidity and temperature module. Added missing operator methods to the temperature template. Expanded the Thermometer class by adding more default behavior, and rewrote a section. Rewrote some of the Pin class logic to clean it up a bit, as well as drop pin values from int to short (even analog reads are pretty small values). Added macro defines to control more debug messages, which leads to the following. Added more debug messages to classes (including Pin and DHT22, amongst others). Removed excessive includes from some files which no longer required them.

+++ Thermostat/Pin.hpp
@@ -26,7 +26,7 @@ class Pin
 {
   public:
     typedef unsigned int Identifier;
-    typedef int Value;
+    typedef short Value;
 
     enum Mode
     {
